Some small fixes
- rgbGen const now is a float that isn't capped at 1.0
- removed stupid cheap bloom which I accidentally committed
- fixed color space of images on hdr lit maps

+++ codemp/rd-rend2/tr_local.h
@@ -881,7 +881,7 @@ typedef struct {
 	waveForm_t		alphaWave;
 	alphaGen_t		alphaGen;
 
-	byte			constantColor[4];			// for CGEN_CONST and AGEN_CONST
+	float			constantColor[4];			// for CGEN_CONST and AGEN_CONST
 
 	uint32_t		stateBits;					// GLS_xxxx mask
